+++ test/test_insensitive_hash.rb
@@ -76,17 +76,23 @@ class TestInsensitiveHash < Test::Unit::TestCase
 
     # FIXME: test insensitive call with encoder
     h = InsensitiveHash[{'Key with spaces' => 1}]
-    h2 = h.insensitive { |str| str.downcase.gsub ' ', '_' }
+    h2 = h.insensitive :underscore => true
+    h3 = h.insensitive :underscore => false
+    assert_raise(ArgumentError) { h.insensitive :underscore => :wat }
 
     assert_equal 1, h['key with spaces']
     assert_nil      h[:key_with_spaces]
+    assert_equal 1, h3['key with spaces']
+    assert_nil      h3[:key_with_spaces]
     assert_equal 1, h2[:KEY_WITH_SPACES]
     assert_equal 1, h2[:Key_with_spaces]
+
     assert_equal ['Key with spaces'], h.keys
     assert_equal ['Key with spaces'], h2.keys
+    assert_equal ['Key with spaces'], h3.keys
 
     h = { 'A key with spaces' => true }
-    ih = h.insensitive { |s| s.downcase.gsub ' ', '_' }
+    ih = h.insensitive :underscore => true
     assert ih[:a_key_with_spaces]
   end
 
@@ -294,16 +300,54 @@ class TestInsensitiveHash < Test::Unit::TestCase
     assert_raise(KeyError) { h.fetch('D') }
   end
 
-  def test_encoder
+  def test_underscore
     h = InsensitiveHash[{'Key with spaces' => 1}]
 
     assert_equal 1, h['key with spaces']
     assert_nil      h[:key_with_spaces]
 
-    h.encoder = proc { |str| str.downcase.gsub(' ', '_') }
-    assert_equal 1, h[:KEY_WITH_SPACES]
-    assert_equal 1, h[:Key_with_spaces]
-    assert_equal ['Key with spaces'], h.keys
+    10.times do
+      assert_raise(ArgumentError) { h.underscore = 'wat' }
+      assert !h.underscore?
+      h.underscore = true
+      assert h.underscore?
+
+      test_keys = [
+        :KEY_WITH_SPACES,
+        :Key_with_spaces,
+        :key_with_spaces,
+        'key with_spaces',
+        'KEY_WITH spaces'
+      ]
+      
+      test_keys.each do |k|
+        assert_equal 1, h[k]
+      end
+      assert_equal ['Key with spaces'], h.keys
+
+      h.underscore = false
+      assert !h.underscore?
+      assert_equal 1, h['KEY WITH SPACES']
+      test_keys.each do |k|
+        assert_nil h[k]
+      end
+      assert_equal ['Key with spaces'], h.keys
+    end
+  end
+
+  def test_underscore_clash
+    h = {}.insensitive
+    h['hello world'] = 1
+    h['hello_world'] = 2
+    assert_equal 1, h['HELLO WORLD']
+    assert_equal 2, h['HELLO_WORLD']
+    assert_equal ['hello world', 'hello_world'], h.keys
+
+    # FIXME
+    h.underscore = true
+    assert_equal ['hello world', 'hello_world'], h.keys
+    assert_equal 2, h['HELLO WORLD']
+    assert_equal 2, h[:HELLO_WORLD]
   end
 end