+++ CHANGELOG.markdown
@@ -1,5 +1,5 @@
 ### 0.2.2 / 2012/02/13
-* Customizable insensitivity
+* :underscore option added
 
 ### 0.2.1 / 2012/02/10
 * Bug fix: Insensitive `fetch`